changed single hook into multiple hooks, one for each usecase
